Address last suggestions
