+++ packages/ia-zendesk-help-widget/src/assets/loader-icon.ts
