+++ packages/ia-zendesk-help-widget/src/assets/question-icon.ts
